Move some classes constructors into .cpp
