+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/absolute_target_speed.hpp
@@ -15,11 +15,10 @@
 #ifndef OPENSCENARIO_INTERPRETER__SYNTAX__ABSOLUTE_TARGET_SPEED_HPP_
 #define OPENSCENARIO_INTERPRETER__SYNTAX__ABSOLUTE_TARGET_SPEED_HPP_
 
-#include <openscenario_interpreter/procedure.hpp>
-#include <openscenario_interpreter/reader/attribute.hpp>
 #include <openscenario_interpreter/scope.hpp>
+#include <openscenario_interpreter/syntax/double.hpp>
 #include <openscenario_interpreter/syntax/entity_ref.hpp>
-#include <openscenario_interpreter/syntax/rule.hpp>
+#include <pugixml.hpp>
 
 namespace openscenario_interpreter
 {
@@ -36,11 +35,7 @@ struct AbsoluteTargetSpeed
 {
   const Double value;
 
-  template <typename Node>
-  explicit AbsoluteTargetSpeed(const Node & node, Scope & scope)
-  : value(readAttribute<Double>("value", node, scope))
-  {
-  }
+  explicit AbsoluteTargetSpeed(const pugi::xml_node &, Scope &);
 
   auto getCalculateAbsoluteTargetSpeed() const
   {
