+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/action.hpp
@@ -20,7 +20,7 @@
 #include <openscenario_interpreter/syntax/private_action.hpp>
 #include <openscenario_interpreter/syntax/storyboard_element.hpp>
 #include <openscenario_interpreter/syntax/user_defined_action.hpp>
-#include <utility>
+#include <pugixml.hpp>
 
 namespace openscenario_interpreter
 {
@@ -42,18 +42,7 @@ struct Action : public Scope, public StoryboardElement<Action>, public Element
 {
   bool overridden = false;
 
-  template <typename Node>
-  explicit Action(const Node & node, Scope & scope)
-  // clang-format off
-  : Scope(scope.makeChildScope(readAttribute<String>("name", node, scope))),
-    Element(
-      choice(node,
-        std::make_pair(     "GlobalAction", [this](auto && node) { return make<     GlobalAction>(node, localScope()); }),
-        std::make_pair("UserDefinedAction", [this](auto && node) { return make<UserDefinedAction>(node, localScope()); }),
-        std::make_pair(    "PrivateAction", [this](auto && node) { return make<    PrivateAction>(node, localScope()); })))
-  // clang-format on
-  {
-  }
+  explicit Action(const pugi::xml_node &, Scope &);
 
   using StoryboardElement::currentState;
 
@@ -72,15 +61,15 @@ struct Action : public Scope, public StoryboardElement<Action>, public Element
 
   using StoryboardElement::evaluate;
 
-  auto ready() const -> bool;
+  /*  */ auto ready() const -> bool;
 
-  auto run() -> void;
+  /*  */ auto run() -> void;
 
-  auto start() -> void;
+  /*  */ auto start() -> void;
 
-  auto stop() -> void;
+  /*  */ auto stop() -> void;
 
-  static constexpr auto stopTriggered() noexcept { return false; }
+  static auto stopTriggered() noexcept -> bool;
 };
 
 auto operator<<(nlohmann::json &, const Action &) -> nlohmann::json &;
