+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/maneuver_group.hpp
@@ -19,6 +19,7 @@
 #include <openscenario_interpreter/syntax/actors.hpp>
 #include <openscenario_interpreter/syntax/maneuver.hpp>
 #include <openscenario_interpreter/syntax/storyboard_element.hpp>
+#include <pugixml.hpp>
 
 namespace openscenario_interpreter
 {
@@ -41,22 +42,7 @@ struct ManeuverGroup : public Scope, public StoryboardElement<ManeuverGroup>, pu
 {
   const Actors actors;
 
-  template <typename Node>
-  explicit ManeuverGroup(const Node & node, Scope & outer_scope)
-  : Scope(outer_scope.makeChildScope(readAttribute<String>("name", node, outer_scope))),
-    StoryboardElement(readAttribute<UnsignedInteger>(
-      "maximumExecutionCount", node, localScope(), UnsignedInteger())),
-    actors(readElement<Actors>("Actors", node, localScope()))
-  {
-    callWithElements(node, "CatalogReference", 0, unbounded, [&](auto && node) {
-      throw UNSUPPORTED_ELEMENT_SPECIFIED(node.name());
-      return unspecified;
-    });
-
-    callWithElements(node, "Maneuver", 0, unbounded, [&](auto && node) {
-      return push_back(readStoryboardElement<Maneuver>(node, localScope()));
-    });
-  }
+  explicit ManeuverGroup(const pugi::xml_node &, Scope &);
 
   using StoryboardElement::evaluate;
