+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/story.hpp
@@ -16,9 +16,9 @@
 #define OPENSCENARIO_INTERPRETER__SYNTAX__STORY_HPP_
 
 #include <nlohmann/json.hpp>
-#include <openscenario_interpreter/syntax/act.hpp>
+#include <openscenario_interpreter/scope.hpp>
 #include <openscenario_interpreter/syntax/storyboard_element.hpp>
-#include <openscenario_interpreter/syntax/string.hpp>
+#include <pugixml.hpp>
 
 namespace openscenario_interpreter
 {
@@ -37,18 +37,7 @@ inline namespace syntax
  * -------------------------------------------------------------------------- */
 struct Story : public Scope, public StoryboardElement<Story>, public Elements
 {
-  template <typename Node>
-  explicit Story(const Node & node, Scope & scope)
-  : Scope(scope.makeChildScope(readAttribute<String>("name", node, scope)))
-  {
-    callWithElements(node, "ParameterDeclarations", 0, 1, [&](auto && node) {
-      return make<ParameterDeclarations>(node, localScope());
-    });
-
-    callWithElements(node, "Act", 1, unbounded, [&](auto && node) {
-      return push_back(readStoryboardElement<Act>(node, localScope()));
-    });
-  }
+  explicit Story(const pugi::xml_node &, Scope &);
 
   using StoryboardElement::evaluate;
