+++ openscenario/openscenario_interpreter/src/syntax/condition.cpp
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <openscenario_interpreter/syntax/by_entity_condition.hpp>
+#include <openscenario_interpreter/syntax/by_value_condition.hpp>
 #include <openscenario_interpreter/syntax/condition.hpp>
 #include <openscenario_interpreter/utility/demangle.hpp>
 
@@ -22,6 +24,20 @@ inline namespace syntax
 static_assert(std::is_standard_layout<ConditionEdge>::value, "");
 
 static_assert(std::is_trivial<ConditionEdge>::value, "");
+
+Condition::Condition(const XML & node, Scope & scope)
+// clang-format off
+: Element(
+    choice(node,
+      std::make_pair("ByEntityCondition", [&](auto && node) { return make<ByEntityCondition>(node, scope); }),
+      std::make_pair( "ByValueCondition", [&](auto && node) { return make< ByValueCondition>(node, scope); }))),
+  name(readAttribute<String>("name", node, scope)),
+  delay(readAttribute<Double>("delay", node, scope, Double())),
+  condition_edge(readAttribute<ConditionEdge>("conditionEdge", node, scope)),
+  current_value(false)
+// clang-format on
+{
+}
 
 auto Condition::evaluate() -> Element
 {
